+++ molecule/python/molecule_api/models/token_update_payload.py
@@ -39,8 +39,8 @@ class TokenUpdatePayload(object):
         'name': 'str',
         'nucleus_model_id': 'str',
         'owner_wallet_id': 'str',
-        'total_supply': 'int',
-        'circulating_supply': 'int',
+        'total_supply': 'float',
+        'circulating_supply': 'float',
         'conract_address': 'str',
         'crowdsale_address': 'str',
         'restrictions': 'TokenRestrictionsPayload',
@@ -205,7 +205,7 @@ class TokenUpdatePayload(object):
         The total supply of the security token  # noqa: E501
 
         :return: The total_supply of this TokenUpdatePayload.  # noqa: E501
-        :rtype: int
+        :rtype: float
         """
         return self._total_supply
 
@@ -216,7 +216,7 @@ class TokenUpdatePayload(object):
         The total supply of the security token  # noqa: E501
 
         :param total_supply: The total_supply of this TokenUpdatePayload.  # noqa: E501
-        :type: int
+        :type: float
         """
 
         self._total_supply = total_supply
@@ -228,7 +228,7 @@ class TokenUpdatePayload(object):
         The amount of tokens in circulation. Defaults to 0  # noqa: E501
 
         :return: The circulating_supply of this TokenUpdatePayload.  # noqa: E501
-        :rtype: int
+        :rtype: float
         """
         return self._circulating_supply
 
@@ -239,7 +239,7 @@ class TokenUpdatePayload(object):
         The amount of tokens in circulation. Defaults to 0  # noqa: E501
 
         :param circulating_supply: The circulating_supply of this TokenUpdatePayload.  # noqa: E501
-        :type: int
+        :type: float
         """
 
         self._circulating_supply = circulating_supply
